Fix deadlock in edge block update due to mistaken assumption.
When incrementing numEdges for an edge block, usually a 0->1 transition means that we are adding the first edge to a brand-new block.
So the old code wrote to smallStamp without synchronizing, assuming that no other thread could reach the block.
But a 0->1 transition for numEdges could also mean that all edges were deleted from the block previously.
In this case we may write to smallStamp inside another thread's critical section, causing a deadlock.

There's actually no need for a special case here. new_eb() initializes smallStamp/largeStamp to INT64_MAX/INT64_MIN,
so the generic code for updating the timestamps will work even when the block is new.