New functions for fast batch insertion of edges.
Summary:

Provides a new family of functions for inserting an entire batch of
edge updates into stinger at once. Because they perform multiple
updates for each edge list traversal, these functions can be much
faster (~10x) than the current implementation, which traverses the
edge list at least once for each update.

Algorithm Overview:

1. Sort the list of edge updates by source vertex ID
2. Each thread grabs a sub-list of updates for a single source
   vertex ID and walks that vertex's edge list
3. For each edge encountered, it performs a binary search on its
   sublist for any matching updates. Empty slots are filled with
   the next available edge in the list.

API:

The new functions are prefixed with `stinger_batch_`, and their
functionality corresponds to the existing functions for inserting
edges:
- stinger_insert_edge -> stinger_batch_insert_edges
- stinger_incr_edge -> stinger_batch_incr_edges
- stinger_insert_edge_pair -> stinger_batch_insert_edge_pairs
- stinger_incr_edge -> stinger_batch_incr_edge_pairs

The functions accept a pair of RandomAccessIterators to the range
of updates to perform, which will be sorted in-place. The caller
must also supply a set of adapter functions for getting/setting the
source, destination, weight, type, and result code for an edge update
as a template parameter. This design allows the functions to be
independent of the underlying edge update type (e.g. EdgeInsertion or
stinger_edge_update)