Enforce alignment of stinger sub-structures.
STINGER makes one large contiguous allocation, then packs several sub-
structures into this chunk of memory. The alignment of each struct
within the chunk depends on the total length of all the structs and
arrays before it.

According to section 8.1.1 of the Intel® 64 and IA-32 Architectures
Developer's Manual: Vol. 3A, "Accesses to cacheable memory that are
split across cache lines and page boundaries are not guaranteed to be
atomic". Operations like readfe() and writeef() can cause deadlocks
if they operate on pointers to unaligned 64-bit data.

Most sub-structures are composed of int64_t's, so they are aligned
regardless of size. But stinger_names includes char arrays, the
size of which is specified by the configuration variable
STINGER_NAME_STR_MAX, plus one for a null terminator byte. The
default value of 255 works, but changing this value can break the
alignment of the other structures, causing deadlocks.

This patch adds several checks to ensure that the all of the
STINGER sub-structures are aligned to an 8-byte boundary.